ICU-13275 Merge warning cleanup fixes into trunk
X-SVN-Rev: 40283

+++ icu4c/source/i18n/plurrule.cpp
@@ -1665,7 +1665,7 @@ const char *PluralAvailableLocalesEnumeration::next(int32_t *resultLength, UErro
     }
     const char *result = ures_getKey(fRes);
     if (resultLength != NULL) {
-        *resultLength = uprv_strlen(result);
+        *resultLength = static_cast<int32_t>(uprv_strlen(result));
     }
     return result;
 }
